+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Added
 
+- replacing all handlers in a `Try` or `TryStar` with the other kind will change the try to the other kind
 - prescribed slicing for `FunctionDef.decorator_list`
 - prescribed slicing for `AsyncFunctionDef.decorator_list`
 - prescribed slicing for `ClassDef.decorator_list`